+++ SwiftGL/Vec3.swift
@@ -48,18 +48,18 @@ struct Vec3 {
     }
     
     var length2: CFloat {
-    get {
-        return x * x + y * y + z * z
-    }
+        get {
+            return x * x + y * y + z * z
+        }
     }
     
     var length: CFloat {
-    get {
-        return sqrtf(self.length2)
-    }
-    set(l) {
-        self = l * normalize(self)
-    }
+        get {
+            return sqrt(self.length2)
+        }
+        set(l) {
+            self = l * normalize(self)
+        }
     }
     
     // Swizzle (Vec2) Properties
